feat(file-manager): Extract parser and converter for fileId into identifier pkg
