+++ file-manager/pkg/backend/identifier/fileid_parser_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package controller
+package identifier
 
 import (
 	"testing"
