+++ file-manager/pkg/backend/s3/FileDownloader.go
@@ -11,6 +11,7 @@ import (
 	"github.com/minio/minio-go/v7"
 	"github.com/pkg/errors"
 	"github.com/telekom/controlplane/file-manager/pkg/backend"
+	"github.com/telekom/controlplane/file-manager/pkg/backend/identifier"
 	"io"
 )
 
@@ -40,7 +41,7 @@ func (s *S3FileDownloader) DownloadFile(ctx context.Context, fileId string) (*io
 	}
 
 	// Convert fileId to S3 path format
-	s3Path, err := ConvertFileIdToS3Path(fileId)
+	s3Path, err := identifier.ConvertFileIdToPath(fileId)
 	if err != nil {
 		log.Error(err, "Failed to convert fileId to S3 path")
 		return nil, errors.Wrap(err, "failed to convert fileId to S3 path")
